all queries in query folder
