+++ client/src/pages/lab_management/manage_equipment/EquipmentEditor.tsx
@@ -8,7 +8,7 @@ import { useQuery } from "@apollo/client";
 import { GET_TRAINING_MODULES } from "../../../queries/trainingQueries";
 import RequestWrapper from "../../../common/RequestWrapper";
 import styled from "styled-components";
-import GET_ROOMS from "../../../queries/getRooms";
+import GET_ROOMS from "../../../queries/roomQueries";
 import { Equipment } from "./EditEquipmentPage";
 import AttachedModule from "./AttachedModule";
 import ArchiveEquipmentButton from "./ArchiveEquipmentButton";
