+++ client/src/pages/maker/take_quiz/QuizResults.tsx
@@ -3,7 +3,7 @@ import { Grid } from "@mui/material";
 import { useParams } from "react-router-dom";
 import { useImmer } from "use-immer";
 import RequestWrapper from "../../../common/RequestWrapper";
-import { GET_LATEST_SUBMISSION, GET_SUBMISSION, GET_SUBMISSIONS } from "../../../queries/getSubmissions";
+import { GET_LATEST_SUBMISSION, GET_SUBMISSION, GET_SUBMISSIONS } from "../../../queries/quizQueries";
 import { GET_MODULE } from "../../../queries/trainingQueries";
 import { Module } from "../../../types/Quiz";
 import Page from "../../Page";
