+++ client/src/pages/maker/training/TrainingPage.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import Page from "../../Page";
 import Explainer from "./Explainer";
-import { gql, useQuery } from "@apollo/client";
+import { useQuery } from "@apollo/client";
 import RequestWrapper2 from "../../../common/RequestWrapper2";
 import { useCurrentUser } from "../../../common/CurrentUserProvider";
 import {
@@ -10,15 +10,8 @@ import {
 } from "../../../common/TrainingModuleUtils";
 import TrainingModuleRow from "../../../common/TrainingModuleRow";
 import SearchBar, { searchFilter } from "../../../common/SearchBar";
+import {GET_ALL_TRAINING_MODULES} from "../../../queries/trainingQueries";
 
-export const GET_ALL_TRAINING_MODULES = gql`
-  query GetAllTrainingModules {
-    modules {
-      id
-      name
-    }
-  }
-`;
 
 export default function TrainingPage() {
   const { passedModules } = useCurrentUser();
