--- client/src/queries/getRooms.ts
@@ -1,12 +1,0 @@
-import { gql } from "@apollo/client";
-
-const GET_ROOMS = gql`
-  query GetRooms {
-    rooms {
-      id
-      name
-    }
-  }
-`;
-
-export default GET_ROOMS;
